end of session 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -4,6 +4,7 @@ using AuctionService.DTOs;
 using AutoMapper;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -51,6 +52,7 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(
             [FromBody] CreateAuctionDto createAuctionDto
@@ -58,7 +60,7 @@ namespace AuctionService.Controllers
         {
             var auction = _mapper.Map<Auction>(createAuctionDto);
             // TODO: add current user as seller
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -74,6 +76,7 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(
             Guid id,
@@ -86,6 +89,9 @@ namespace AuctionService.Controllers
 
             if (auction == null)
                 return NotFound();
+
+            if (auction.Seller != User.Identity.Name)
+                return Forbid();
             // Logging dữ liệu nhận được
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
@@ -104,6 +110,7 @@ namespace AuctionService.Controllers
             return Ok();
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -111,6 +118,9 @@ namespace AuctionService.Controllers
 
             if (auction == null)
                 return NotFound();
+
+            if (auction.Seller != User.Identity.Name)
+                return Forbid();
 
             _context.Auctions.Remove(auction);
